% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apply_tac}
\alias{apply_tac}
\title{Apply TAC-ABC fitting logic}
\usage{
apply_tac(abc, tacpar, total_oy_cap = 1945)
}
\arguments{
\item{abc}{vector of calculated ABCs for the species in the complex}

\item{tacpar}{the list object from read_tacpar()}

\item{total_oy_cap}{The OY cap (e.g., 2000 for 2 million t)}
}
\description{
Apply TAC-ABC fitting logic
}
