% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_output}
\alias{format_output}
\title{Format Projection Results (Dynamic Variable)}
\usage{
format_output(projection_data, var = "ssb")
}
\arguments{
\item{projection_data}{The raw output from run_projections()}

\item{var}{The variable to summarize: "ssb", "catch", or "f"}
}
\value{
A tidytable matching the ADMB csv format
}
\description{
Format Projection Results (Dynamic Variable)
}
