% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_recruitment}
\alias{get_recruitment}
\title{Calculate recruitment (supports SR models or mean recruitment)}
\usage{
get_recruitment(
  ssb,
  b0,
  h,
  r0,
  r_mean = NULL,
  h_mean = NULL,
  type = "Mean",
  rho = 0,
  sigma_r = 0.5,
  epsilon_prev = 0,
  rec_dev = rnorm(1)
)
}
\arguments{
\item{ssb}{spawning stock biomass (ignored if type = "Mean")}

\item{b0}{unfished equilibrium biomass (ignored if type = "Mean")}

\item{h}{steepness (ignored if type = "Mean")}

\item{r0}{unfished recruitment (used as parameter for SR curves)}

\item{r_mean}{average recruitment (used ONLY if type = "Mean")}

\item{h_mean}{harmonic mean recruitment (used ONLY if type = "IG" (Inverse Gaussian))}

\item{type}{"BH", "Ricker", "Mean", "or "IG"}

\item{rho}{autocorrelation coefficient}

\item{sigma_r}{log-scale standard deviation of recruitment}

\item{epsilon_prev}{the total log-deviation from the previous year}

\item{rec_dev}{current year random normal deviate}
}
\value{
a list containing calculated recruitment and the error term (chi)
}
\description{
Calculate recruitment (supports SR models or mean recruitment)
}
