% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_tier_f}
\alias{get_tier_f}
\title{Calculate F based on NPFMC Tier System (Amendment 56)}
\usage{
get_tier_f(ssb, f_ref, b_ref, alpha = 0.05, ssl_protection = FALSE)
}
\arguments{
\item{ssb}{current spawning biomass}

\item{f_ref}{reference F (F40, F35, or Fabc)}

\item{b_ref}{reference biomass (B40, B35, or B100 depending on tier)}

\item{alpha}{slope cut-off (usually 0.05)}

\item{ssl_protection}{boolean. if TRUE, F cuts to 0 at higher B/Bref ratio (0.2)}
}
\value{
the target Fishing Mortality (F)
}
\description{
Calculate F based on NPFMC Tier System (Amendment 56)
}
