% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inverse_gaussian}
\alias{inverse_gaussian}
\title{Inverse Gaussian Random Variable (ADMB style)}
\usage{
inverse_gaussian(n, arithmetic_mean, harmonic_mean)
}
\arguments{
\item{n}{number of deviates to generate}

\item{arithmetic_mean}{the standard mean of historical recruitment}

\item{harmonic_mean}{the harmonic mean of historical recruitment}
}
\value{
vector of recruitment deviates
}
\description{
Inverse Gaussian Random Variable (ADMB style)
}
