% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m22.1}
\alias{m22.1}
\title{example RTMButils dataset}
\format{
A data list with assessment outputs
}
\usage{
m22.1
}
\description{
Note this is old and may no longer work - haven't checked...
}
\keyword{datasets}
