% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proj_5yr_table}
\alias{proj_5yr_table}
\title{Generate the 5-Year Expected Catch and Projection Table}
\usage{
proj_5yr_table(raw_proj, report, start_year)
}
\arguments{
\item{raw_proj}{The raw output table from run_projections()}

\item{report}{The RTMB report object (needed for B100\% / B0)}

\item{start_year}{Current assessment year}
}
\value{
A tidytable formatted for the 5-year projection table
}
\description{
Generate the 5-Year Expected Catch and Projection Table
}
