% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proj_rtmb}
\alias{proj_rtmb}
\title{Generate Executive Summary and Projections}
\usage{
proj_rtmb(
  report,
  year,
  future_catch,
  yield_ratio,
  output_dir = "processed",
  perc_max = 0.5
)
}
\arguments{
\item{report}{The RTMB report object}

\item{year}{current assessment year}

\item{future_catch}{numeric vector of catches for the next 2 years}

\item{yield_ratio}{ratio to scale maxABC to Author ABC (e.g. 0.8)}

\item{output_dir}{where to save the CSVs}

\item{perc_max}{allow for a maximum percent of F to be used in scenario 4 - between 0-1}
}
\description{
Generate Executive Summary and Projections
}
