% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{project_step}
\alias{project_step}
\title{Single Year Population Projection Step}
\usage{
project_step(
  n_at_age,
  m_at_age,
  sel_at_age,
  wt_at_age,
  mat_at_age,
  f_current,
  rec_params,
  prev_chi,
  spawn_frac
)
}
\arguments{
\item{n_at_age}{vector of abundance at age at start of year t}

\item{m_at_age}{vector of M at age}

\item{sel_at_age}{vector of Selectivity at age}

\item{wt_at_age}{vector of Weight at age}

\item{mat_at_age}{vector of Maturity at age}

\item{f_current}{scalar Fishing mortality for year t}

\item{rec_params}{list of SR parameters (b0, h, r0, rho, sigma_r)}

\item{prev_chi}{the error term from year t-1 (for Rho calc)}

\item{spawn_frac}{fraction of year elapsed before spawning}
}
\value{
A list containing N_at_age (t+1), Catch (t), SSB (t), and new chi
}
\description{
Single Year Population Projection Step
}
