% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_projections}
\alias{run_projections}
\title{Run NPFMC Scenarios 1-7}
\usage{
run_projections(
  report,
  future_catch = NULL,
  yield_ratio = NULL,
  n_sims = 1500,
  n_years = 14,
  unit_conversion = 1,
  sex_ratio = 0.5,
  rec_model = "IG",
  sigma_r_override = NULL,
  scenarios = 1:7,
  perc_max,
  seed = 123
)
}
\arguments{
\item{report}{The RTMB report object}

\item{future_catch}{vector of catches (e.g., c("2024" = 1170, "2025" = 1968))}

\item{yield_ratio}{A numeric value (e.g., 0.3598).
If provided, Scenario 2 (Author F) catches for years 2 & 3 will be calculated as:
Scenario 1 Catch * yield_ratio.}

\item{n_sims}{number of simulations, default 1500}

\item{n_years}{number of projection years}

\item{unit_conversion}{Scaling factor to align Biomass units with Catch units.
\itemize{
\item If Nat=Millions, Waa=Grams, Catch=Tons -> Use 1.0
\item If Nat=Thousands, Waa=Grams, Catch=Tons -> Use 0.001
}}

\item{sex_ratio}{proportion of females for SSB calculation (default 0.5)}

\item{rec_model}{the recruitment type: "Mean", "IG" (mean using Inverse Gaussian),
"Ricker", or "BH" (Beverton-Holt).}

\item{sigma_r_override}{optional numeric to override report$sigmaR.}

\item{scenarios}{the NPFMC scenarios to run (1:7)}

\item{perc_max}{allow for a maximum percent of F to be used in scenario 4 - between 0-1}

\item{seed}{the seed you want default: 123}
}
\description{
Run NPFMC Scenarios 1-7
}
