% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_f_from_catch}
\alias{get_f_from_catch}
\title{Solve for F given a Target Catch}
\usage{
get_f_from_catch(
  target_catch,
  n_at_age,
  m_at_age,
  sel_at_age,
  wt_at_age,
  scalar = 1
)
}
\description{
Solve for F given a Target Catch
}
